server/audio-service/requirements.txt
fastapi
uvicorn[standard]
python-multipart
openai-whisper
torch
webrtcvad
speechbrain
simple-diarizer
python-dotenv